requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
backports-datetime-fromisoformat==2.0.3
backports.tarfile==1.2.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
email-validator==2.3.0
essentia==2.1b6.dev1177
essentia-tensorflow==2.1b6.dev1389
et_xmlfile==2.0.0
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
faiss-cpu==1.11.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-KVSession==0.6.2
flask-marshmallow==1.3.0
Flask-PyMongo==3.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.30.2
id==1.5.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.24
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.38
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
mongoengine==0.29.1
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
namex==0.0.9
narwhals==1.42.1
networkx==3.4.2
nh3==0.3.0
numba==0.61.2
numpy==1.26.4
openai==1.76.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.8
pooch==1.8.2
proglog==0.1.12
protobuf==4.25.8
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyformgen==0.2.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readme_renderer==44.0
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.33.0
sentence-transformers==4.1.0
sentry-sdk==2.46.0
shellingham==1.5.4
simplekv==0.14.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
spotipy==2.24.0
SQLAlchemy==2.0.40
starlette==0.50.0
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
twine==6.1.0
typer==0.20.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yt-dlp==2025.6.30
ytmusicapi==1.10.3
zipp==3.23.0
zstandard==0.23.0
